ref: add tests file, convert cw721 errors into our errors
